+++ src/engine/eval/tapered_eval.rs
@@ -5,21 +5,22 @@ use crate::engine::eval::WhiteEval;
 
 /// A midgame and endgame evaluation
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
-pub struct PhasedEval(i16, i16);
+pub struct PhasedEval(i32);
 
 impl PhasedEval {
-    pub const ZERO: Self = Self(0, 0);
+    pub const ZERO: Self = Self(0);
 
     pub const fn new(midgame: i16, endgame: i16) -> Self {
-        Self(midgame, endgame)
+        Self(((endgame as i32) << 16) + midgame as i32)
     }
 
+    #[allow(clippy::cast_possible_truncation)]
     pub fn midgame(self) -> WhiteEval {
-        WhiteEval(self.0)
+        WhiteEval(self.0 as i16)
     }
 
     pub fn endgame(self) -> WhiteEval {
-        WhiteEval(self.1)
+        WhiteEval(((self.0 + 0x8000) >> 16) as i16)
     }
 }
 
@@ -27,14 +28,13 @@ impl std::ops::Add for PhasedEval {
     type Output = Self;
 
     fn add(self, rhs: Self) -> Self::Output {
-        Self(self.0 + rhs.0, self.1 + rhs.1)
+        Self(self.0 + rhs.0)
     }
 }
 
 impl std::ops::AddAssign for PhasedEval {
     fn add_assign(&mut self, rhs: Self) {
         self.0 += rhs.0;
-        self.1 += rhs.1;
     }
 }
 
@@ -42,22 +42,13 @@ impl std::ops::Sub for PhasedEval {
     type Output = Self;
 
     fn sub(self, rhs: Self) -> Self::Output {
-        Self(self.0 - rhs.0, self.1 - rhs.1)
+        Self(self.0 - rhs.0)
     }
 }
 
 impl std::ops::SubAssign for PhasedEval {
     fn sub_assign(&mut self, rhs: Self) {
         self.0 -= rhs.0;
-        self.1 -= rhs.1;
-    }
-}
-
-impl std::ops::Mul<i16> for PhasedEval {
-    type Output = Self;
-
-    fn mul(self, rhs: i16) -> Self::Output {
-        Self(self.0 * rhs, self.1 * rhs)
     }
 }
 
@@ -65,7 +56,7 @@ impl std::ops::Neg for PhasedEval {
     type Output = Self;
 
     fn neg(self) -> Self::Output {
-        Self(self.0.saturating_neg(), self.1.saturating_neg())
+        Self(self.0.saturating_neg())
     }
 }
 
@@ -103,8 +94,8 @@ pub fn taper(phase_value: i16, eval: PhasedEval) -> WhiteEval {
     let midgame_phase_value = phase_value.min(PHASE_COUNT_MAX);
     let endgame_phase_value = PHASE_COUNT_MAX - phase_value;
 
-    let midgame_eval = i64::from(eval.0);
-    let endgame_eval = i64::from(eval.1);
+    let midgame_eval = i64::from(eval.midgame().0);
+    let endgame_eval = i64::from(eval.endgame().0);
 
     let eval = (midgame_eval * midgame_phase_value + endgame_eval * endgame_phase_value) / 24;
     WhiteEval(i16::try_from(eval).unwrap())
@@ -116,10 +107,18 @@ mod tests {
 
     #[test]
     fn test_eval_calculation_does_not_overflow() {
-        let phased_eval = PhasedEval(2456, 2393);
+        let phased_eval = PhasedEval::new(2456, 2393);
         let phase_value = 9;
 
         let eval = taper(phase_value, phased_eval);
         assert!(eval > WhiteEval(0));
+    }
+
+    #[test]
+    fn test_eval_retains_midgame_and_endgame_vals() {
+        let phased_eval = PhasedEval::new(2456, 2393);
+
+        assert_eq!(phased_eval.midgame().0, 2456);
+        assert_eq!(phased_eval.endgame().0, 2393);
     }
 }